+++ resources/views/components/layouts/app.blade.php
@@ -14,7 +14,7 @@
 
     </head>
 
-    <body data-bs-theme="{{ Auth::user()->view_mode ? "dark" : 'light'}}">
+    <body data-bs-theme="{{ ( isset(Auth::user()->view_mode) && Auth::user()->view_mode ? "dark" : 'light') ?? 'light'}}">
         <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
         <header>
             @livewire( Navbar::Class )