+++ resources/views/livewire/dashboard.blade.php
@@ -1,3 +1,7 @@
 <div>
-    <h1> Dashboard. </h1>
+    @can('permissions_index')
+        <h1> Dashboard. </h1>
+    @else
+            <livewire:UnauthorizedPage/>
+    @endcan
 </div>