+++ CHANGELOG.md
@@ -17,6 +17,7 @@ php artisan view:clear -->
 
 ## [0.0.6] - 2026-08-24
 ### Added 
+- Vista de Contacto envia emails según array en constants.php
 ### Changed
 - Register, mensajes de error y navbar vista clientes Fixed